Struct swiftlet_quic::endpoint::Endpoint

pub struct Endpoint { /* private fields */ }

The Quic Endpoint structure

Implementations

impl Endpoint

pub fn new_server( bind_addr: SocketAddr, alpn: &[u8], cert_path: &str, pkey_path: &str, config: Config ) -> Result<Self, Error>

Create a QUIC Server Endpoint

pub fn new_client( bind_addr: SocketAddr, alpn: &[u8], cert_path: &str, config: Config ) -> Result<Self, Error>

Create a QUIC Client Endpoint

pub fn add_client_connection( &mut self, peer_addr: SocketAddr, server_name: &str ) -> Result<(), Error>

Add a connection for a Client Endpoint

Must be used on a Client and not a Server otherwise an error will be thrown

pub fn new_client_with_first_connection( bind_addr: SocketAddr, alpn: &[u8], cert_path: &str, peer_addr: SocketAddr, server_name: &str, config: Config ) -> Result<Self, Error>

Create a QUIC Client Endpoint with an initial connection

pub fn get_num_connections(&self) -> usize

Get the number of connections that the Endpoint is managing

pub fn update_keep_alive_duration(&mut self, duration_opt: Option<Duration>)

Update the keep alive duration time

Will disable the keep alive functionality if set to None

pub fn close_connection( &mut self, cid: &ConnectionId, error_code: u64 ) -> Result<bool, Error>

Close a connection with a given error code value

Returns true when connection close process has started

pub fn main_stream_send( &mut self, cid: &ConnectionId, send_data: Vec<u8> ) -> Result<(), Error>

Send data over the main stream

A reminder that the Endpoint connection will be taking ownership of the data so it can be sent out when possible

pub fn background_stream_send( &mut self, cid: &ConnectionId, send_data: Vec<u8> ) -> Result<(), Error>

Send data over the background stream

A reminder that the Endpoint connection will be taking ownership of the data so it can be sent out when possible